Extract minimal archive handlimg from test #2233
This is to plan use this in rubygems
